fix(Navigation): segment components in sub components
